@@ -1,6 +1,6 @@
 {
   "name": "choerodon-front-boot",
-  "version": "0.7.0",
+  "version": "0.7.1",
   "description": "Tools to build website for Choerodon",
   "main": "lib/containers/components/index.js",
   "bin": {
@@ -11,17 +11,6 @@
     "bin",
     "structure"
   ],
-  "scripts": {
-    "prestart": "npm run compile",
-    "start": "node ./bin/choerodon-front-boot start --config ./choerodon.config.js",
-    "prebuild": "npm run compile",
-    "build": "node ./bin/choerodon-front-boot build --config ./choerodon.config.js",
-    "clean": "rimraf lib",
-    "watch": "gulp watch",
-    "compile": "gulp compile",
-    "prepublish": "npm run compile",
-    "postpublish": "npm run clean"
-  },
   "author": "Hugh <99571603@qq.com>",
   "license": "ISC",
   "dependencies": {
@@ -29,7 +18,7 @@
     "autoprefixer": "^6.7.7",
     "axios": "^0.16.2",
     "babel-core": "^6.25.0",
-    "babel-eslint": "^7.2.3",
+    "babel-eslint": "^8.2.6",
     "babel-loader": "^6.4.1",
     "babel-plugin-import": "^1.2.1",
     "babel-plugin-lodash": "^3.2.11",
@@ -51,19 +40,19 @@
     "case-sensitive-paths-webpack-plugin": "^2.1.2",
     "chalk": "^2.4.1",
     "change-case": "^3.0.1",
-    "choerodon-ui": "0.3.5",
+    "choerodon-ui": "0.3.6",
     "classnames": "^2.2.5",
     "commander": "^2.9.0",
     "component-classes": "^1.x",
     "css-loader": "^0.25.0",
-    "eslint": "^4.2.0",
-    "eslint-config-airbnb": "^15.1.0",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.5.1",
+    "eslint": "^5.3.0",
+    "eslint-config-airbnb": "latest",
+    "eslint-plugin-babel": "^5.1.0",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-react": "^7.10.0",
     "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.9.0",
+    "file-loader": "^1.1.6",
     "friendly-errors-webpack-plugin": "^1.7.0",
     "glob": "^7.1.2",
     "history": "^4.6.3",
@@ -71,6 +60,7 @@
     "json-loader": "^0.5.7",
     "less": "^2.7.2",
     "less-loader": "^3.0.0",
+    "lint-staged": "^7.2.2",
     "lodash": "^4.17.4",
     "merge-stream": "^1.0.1",
     "mkdirp": "^0.5.1",
@@ -81,6 +71,7 @@
     "nunjucks": "^3.1.3",
     "object.omit": "^3.0.0",
     "postcss-loader": "^2.1.5",
+    "pre-commit": "^1.2.2",
     "prop-types": "^15.5.10",
     "query-string": "^4.1.0",
     "react": "^16.4.1",
@@ -95,9 +86,6 @@
     "rxjs": "^5.5.6",
     "sass-loader": "^6.0.6",
     "style-loader": "0.13.1",
-    "stylelint": "^8.4.0",
-    "stylelint-config-standard": "^18.0.0",
-    "stylelint-webpack-plugin": "^0.10.1",
     "ts-loader": "^2.0.2",
     "ts-node": "^4.0.2",
     "tslint": "^5.9.1",
@@ -107,6 +95,7 @@
     "universal-cookie": "^2.0.8",
     "url": "^0.11.0",
     "url-loader": "^1.0.1",
+    "warning": "^3.0.0",
     "webpack": "^3.12.0",
     "webpack-dev-server": "^2.9.1"
   },
@@ -116,5 +105,30 @@
     "gulp-babel": "^7.0.1",
     "merge2": "^1.2.2",
     "through2": "^2.0.3"
-  }
+  },
+  "scripts": {
+    "prestart": "npm run compile",
+    "start": "node ./bin/choerodon-front-boot start --config ./choerodon.config.js",
+    "prebuild": "npm run compile",
+    "build": "node ./bin/choerodon-front-boot build --config ./choerodon.config.js",
+    "lint": "npm run lint:es && npm run lint:style",
+    "lint:es": "eslint src/**/*.{js,jsx} test ./choerodon.config.js",
+    "lint:style": "stylelint src/**/*.scss --syntax scss",
+    "lint-staged": "lint-staged",
+    "lint-staged:es": "eslint test ./choerodon.config.js",
+    "clean": "rimraf lib",
+    "watch": "gulp watch",
+    "compile": "gulp compile",
+    "prepublish": "npm run compile",
+    "postpublish": "npm run clean"
+  },
+  "lint-staged": {
+    "src/**/*.{js,jsx}": [
+      "npm run lint-staged:es"
+    ],
+    "src/**/*.scss": "stylelint --syntax scss"
+  },
+  "pre-commit": [
+    "lint-staged"
+  ]
 }
