@@ -1,4 +1,3 @@
-/*eslint-disable*/
 import React, { Component } from 'react';
 import { addLocaleData, IntlProvider } from 'react-intl';
 import { Observable } from 'rxjs/Observable';
@@ -12,9 +11,10 @@ import 'rxjs/add/observable/fromPromise';
 
 export default function asyncLocaleProvider(locale, getMessage, getLocaleData) {
   return class AsyncLocaleProvider extends Component {
-
     static Message = null;
+
     static needLoadLocaleData = true && getLocaleData;
+
     static LocaleData = null;
 
     state = {
@@ -27,58 +27,55 @@ export default function asyncLocaleProvider(locale, getMessage, getLocaleData) {
       const { Message, LocaleData, needLoadLocaleData } = this.state;
 
       if (!Message || (needLoadLocaleData && !LocaleData)) {
-        this._componentWillUnmountSubject = new Subject();
+        this.componentWillUnmountSubject = new Subject();
 
         const streams = [
           Message
             ? Observable.of(Message)
-              .takeUntil(this._componentWillUnmountSubject)
+              .takeUntil(this.componentWillUnmountSubject)
             : Observable.fromPromise(getMessage())
               .map(esModule)
-              .map(Message => {
-                AsyncLocaleProvider.Message = Message;
-                return Message;
+              .map((message) => {
+                AsyncLocaleProvider.Message = message;
+                return message;
               })
-              .takeUntil(this._componentWillUnmountSubject),
+              .takeUntil(this.componentWillUnmountSubject),
         ];
 
         if (getLocaleData) {
           streams.push(
             needLoadLocaleData && LocaleData
               ? Observable.of(LocaleData)
-                .takeUntil(this._componentWillUnmountSubject)
+                .takeUntil(this.componentWillUnmountSubject)
               : Observable.fromPromise(getLocaleData())
                 .map(esModule)
-                .map(LocaleData => {
-                  return LocaleData;
-                })
-                .takeUntil(this._componentWillUnmountSubject),
+                .takeUntil(this.componentWillUnmountSubject),
           );
         }
 
         Observable.zip(...streams)
-          .takeUntil(this._componentWillUnmountSubject)
-          .subscribe(([Message, LocaleData]) => {
-            if (this._mounted) {
-              this.setState({ Message, LocaleData });
+          .takeUntil(this.componentWillUnmountSubject)
+          .subscribe(([message, localeData]) => {
+            if (this.mounted) {
+              this.setState({ Message: message, LocaleData: localeData });
             } else {
-              this.state.Message = Message;
-              this.state.LocaleData = LocaleData;
+              this.state.Message = message;
+              this.state.LocaleData = localeData;
             }
 
-            this._componentWillUnmountSubject.unsubscribe();
+            this.componentWillUnmountSubject.unsubscribe();
           });
       }
     }
 
     componentDidMount() {
-      this._mounted = true;
+      this.mounted = true;
     }
 
     componentWillUnmount() {
-      if (this._componentWillUnmountSubject && !this._componentWillUnmountSubject.closed) {
-        this._componentWillUnmountSubject.next();
-        this._componentWillUnmountSubject.unsubscribe();
+      if (this.componentWillUnmountSubject && !this.componentWillUnmountSubject.closed) {
+        this.componentWillUnmountSubject.next();
+        this.componentWillUnmountSubject.unsubscribe();
       }
     }
 
@@ -90,7 +87,6 @@ export default function asyncLocaleProvider(locale, getMessage, getLocaleData) {
         } else {
           addLocaleData(LocaleData);
         }
-
       }
       return Message ? <IntlProvider {...this.props} locale={locale.replace('_', '-')} messages={Message} /> : null;
     }
